// libs/ui/core/src/lib/message/message-types.ts
export enum MessageType {
  Info = 'info',
  Error = 'error',
  Loading = 'loading',
  Success = 'success',
  Warning = 'warning',
}

export enum MessagePlacement {
  TopRight = 'top-right',
  Top = 'top',
  TopLeft = 'top-left',
  BottomLeft = 'bottom-left',
  Bottom = 'bottom',
  BottomRight = 'bottom-right',
}